Sorted mode for xdigit tool
The xdigit-sorted is implemented as special mode for xdigit. In case the
input to xdigit is marked --sorted, it'll use a faster non-xawk
algorithm to calculate the class labels. This is done by readin both
DATA and BINS line by line. This is reasonably faster as it doesn't have
to check all classes for each input value.